+++ app/src/main/java/com/samueva/remindme/NotificationPublisher.java
@@ -11,8 +11,8 @@ public class NotificationPublisher extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
         NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-        long taskId = intent.getIntExtra("task_id", 0);
+        int taskId = intent.getIntExtra("task_id", 0);
         Notification notification = intent.getParcelableExtra("notification");
-        notificationManager.notify((int) taskId, notification);
+        notificationManager.notify(taskId, notification);
     }
 }
